pyproject.toml
[project]
name = "img-gen"
version = "0.1.0"
description = "Image generation using Google Gemini / Imagen API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.63.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
]
